requirements.txt
psutil==5.9.8
numpy==1.26.4
requests==2.31.0
google-generativeai==0.8.3
